requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai>=1.6.1
pinecone>=4.0.0
requests>=2.31.0
PyPDF2>=3.0.1
python-multipart>=0.0.6